requirements.txt
Flask==2.3.0
werkzeug==2.3.7
Flask-SQLAlchemy==3.0.2
pytest==7.4.0
flask_cors==4.0.0
flask_migrate==4.0.4
flask_restful==0.3.10
python-dotenv==0.15.0
pytest-dotenv==0.5.2
psycopg2==2.9.7
markupsafe==2.1.1
